refactor(app): fix project structure
