+++ apps/codeimage/src/App.tsx
@@ -14,7 +14,6 @@ import {useI18n} from '@codeimage/locale';
 import {useModality} from './core/hooks/isMobile';
 import {BottomBar} from './components/BottomBar/BottomBar';
 import {FrameHandler} from './components/Frame/FrameHandler';
-import {EditorSidebar} from './components/LeftSidebar/EditorSidebar';
 import {NotificationHandler} from './ui/Toast/SnackbarHost';
 import ReloadPrompt from './components/PromptUpdate/PromptUpdate';
 import {PortalHost} from './ui/PortalHost/PortalHost';
@@ -26,6 +25,7 @@ import {initEffects} from '@ngneat/effects';
 import {onTabNameChange$, onThemeChange$} from '@codeimage/store/effect';
 import {uiStore} from './state/ui';
 import {useEffects} from './core/store/use-effect';
+import {EditorSidebar} from './components/PropertyEditor/EditorSidebar';
 
 initEffects();
 
@@ -45,8 +45,9 @@ const App = () => {
     <Scaffold>
       <NotificationHandler />
       <ReloadPrompt />
+
       <Show when={modality === 'full'}>
-        <Sidebar>
+        <Sidebar position={'left'}>
           <EditorSidebar />
         </Sidebar>
       </Show>
@@ -93,13 +94,14 @@ const App = () => {
         <Footer />
       </Canvas>
 
-      {modality === 'mobile' ? (
-        <BottomBar portalHostRef={portalHostRef()} />
-      ) : (
-        <Sidebar>
+      <Show
+        when={modality === 'full'}
+        fallback={<BottomBar portalHostRef={portalHostRef()} />}
+      >
+        <Sidebar position={'right'}>
           <ThemeSwitcher orientation={'vertical'} />
         </Sidebar>
-      )}
+      </Show>
     </Scaffold>
   );
 };
